requirements.txt
apache-airflow==2.8.4
apache-airflow-providers-celery==3.7.0
pyodbc==5.1.0
psycopg2==2.9.9
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.14.0
psycopg2-binary==2.9.9
pandas==2.2.2